Implement dynamic loading of the menu page
The final part of this project required a lot of DOM manipulation with JavaScript.
To aid with this, I created a function that created item cards, which I could append easily to the parent container by calling the function and supplying three parameters.

Formatting-wise, I opted to use CSS Grid for the main container, but within each item card itself I used flexbox to arrange the elements.
This marks the end of this project, which I will attempt to host on GitHub pages.